+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.3] - 2019-04-15
 ### Added
 - Added chat console alert for Bingo.
+- Resizeable Board
+
+### Changed
+- Middle click on the minimap icon now resets the board
+- Right click on the minimap icon now opens config menu
 
 ## [0.1.2] - 2019-04-07
 ### Added